# docs/swagger.yaml
definitions:
  models.AccessTokenRes:
    properties:
      accessToken:
        type: string
      expAccessToken:
        type: string
    required:
    - accessToken
    - expAccessToken
    type: object
...
  models.LoginResponse:
    properties:
      accessToken:
        type: string
      expAccessToken:
        type: string
      expRefreshToken:
        type: string
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
  models.UserResponse:
    properties:
      createdAt:
        type: integer
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      isAdmin:
        type: boolean
      lastName:
        type: string
      picture:
        type: string
      updatedAt:
        type: integer
      verifiedEmail:
        type: boolean
    required:
    - firstName
    - lastName
    type: object
...
  swagger.EmptyData:
    type: object
  swagger.FailedLoginResponse:
    properties:
      data:
        $ref: '#/definitions/swagger.EmptyData'
      msg:
        example: No user found with entered credentials
        type: string
      ok:
        example: false
        type: boolean
    type: object
  swagger.FailedResponse:
    properties:
      data:
        $ref: '#/definitions/swagger.EmptyData'
      msg:
        example: Error or warnnig message
        type: string
      ok:
        example: false
        type: boolean
    type: object
  swagger.FailedValidationResponse:
    properties:
      data:
        $ref: '#/definitions/swagger.validationErrors'
      msg:
        example: Please review your entered data
        type: string
      ok:
        example: false
        type: boolean
    type: object
  swagger.LoginResponse:
    properties:
      data:
        $ref: '#/definitions/models.LoginResponse'
      msg:
        example: Successful message
        type: string
      ok:
        example: true
        type: boolean
    type: object
  swagger.NotFoundResponse:
    properties:
      data:
        $ref: '#/definitions/swagger.EmptyData'
      msg:
        example: 404 not found!
        type: string
      ok:
        example: false
        type: boolean
    type: object
...
  swagger.SuccessResponse:
    properties:
      data:
        $ref: '#/definitions/swagger.EmptyData'
      msg:
        example: Successful message
        type: string
      ok:
        example: true
        type: boolean
    type: object
  swagger.SuccessVerifyAccessTokenResponse:
    properties:
      data:
        $ref: '#/definitions/models.AccessTokenRes'
      msg:
        example: Successful message
        type: string
      ok:
        example: true
        type: boolean
    type: object
  swagger.UnauthenticatedResponse:
    properties:
      data:
        $ref: '#/definitions/swagger.EmptyData'
      msg:
        example: You must login first!
        type: string
      ok:
        example: false
        type: boolean
    type: object
...
  swagger.validationErrors:
    properties:
      errors:
        additionalProperties:
          type: string
        example:
          field1: This field is required
          field2: This field must be numeric
        type: object
    type: object
info:
  contact: {}
paths:
...
  /auth/access-token-verify:
    post:
      consumes:
      - application/json
      description: jwt access token verify
      parameters:
      - description: accessToken
        in: query
        name: accessToken
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.FailedResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/swagger.FailedValidationResponse'
      summary: access token verify
      tags:
      - auth
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: forgot password
      parameters:
      - description: unique email
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.FailedResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/swagger.FailedValidationResponse'
      summary: forgot password
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: jwt login
      parameters:
      - description: email
        in: query
        name: email
        required: true
        type: string
      - description: send user operating system + browser name in this param
        in: query
        name: deviceName
        required: true
        type: string
      - description: password
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.FailedLoginResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/swagger.FailedValidationResponse'
      summary: login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: jwt logout , atuhentication required
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthenticatedResponse'
      summary: logout
      tags:
      - auth
  /auth/recover-password:
    post:
      consumes:
      - application/json
      description: Let user change it password with forgot token
      parameters:
      - description: token
        in: query
        name: token
        required: true
        type: string
      - description: 'password that have at least 8 length and contain an alphabet
          and number '
        in: query
        name: password
        required: true
        type: string
      - description: 'repeatPassword that have at least 8 length and contain an alphabet
          and number '
        in: query
        name: repeatPassword
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/swagger.NotFoundResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/swagger.FailedValidationResponse'
      summary: recover-password
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: jwt register
      parameters:
      - description: unique email
        in: query
        name: email
        required: true
        type: string
      - description: 'password that have at least 8 length and contain an alphabet
          and number '
        in: query
        name: password
        required: true
        type: string
      - description: 'repeatPassword that have at least 8 length and contain an alphabet
          and number '
        in: query
        name: repeatPassword
        required: true
        type: string
      - description: firstName
        in: query
        name: firstName
        required: true
        type: string
      - description: lastName
        in: query
        name: lastName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.SuccessResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/swagger.FailedValidationResponse'
      summary: register
      tags:
      - auth
  /auth/renew-access-token:
    post:
      consumes:
      - application/json
      description: jwt renew access token
      parameters:
      - description: refreshToken
        in: query
        name: refreshToken
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.SuccessVerifyAccessTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.FailedResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/swagger.FailedValidationResponse'
      summary: renew access token
      tags:
      - auth
  /auth/resend-verify-email:
    post:
      consumes:
      - application/json
      description: resend-verify-email
      parameters:
      - description: unique email
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.SuccessResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/swagger.FailedValidationResponse'
      summary: resend-verify-email
      tags:
      - auth
  /auth/verify-email:
    post:
      consumes:
      - application/json
      description: verify-email
      parameters:
      - description: token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.FailedResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/swagger.FailedValidationResponse'
      summary: verify-email
      tags:
      - auth
...
swagger: "2.0"